+++ crates/web-client/js/types/index.d.ts
@@ -23,6 +23,7 @@ export {
   ForeignAccount,
   FungibleAsset,
   FungibleAssetDelta,
+  InputNoteRecord,
   InputNoteState,
   Library,
   NewSwapTransactionResult,